+++ scripts/read_contribute_streams.py
@@ -24,8 +24,8 @@ import json
 import os
 
 import dotenv
-from ceramic import Ceramic
-from streams import (
+from ceramic.ceramic import Ceramic
+from ceramic.streams import (
     CONTRIBUTE_PROD_CENTAURS_STREAM_ID,
     CONTRIBUTE_PROD_DB_STREAM_ID,
     CONTRIBUTE_PROD_MANUAL_STREAM_ID,
